extract homepage reusable jsx code and data constants to independent files.
