+++ components/homepage/HomepageAbout.js
@@ -2,6 +2,7 @@ import React, { useState } from 'react';
 import Swiper from 'react-id-swiper';
 import ModalVideo from 'react-modal-video';
 import 'swiper/css/swiper.css';
+import AboutSectionItem from './HomepageAboutSectionItem';
 
 const HomepageAbout = () => {
   const [swiper, setSwiper] = useState(null);
@@ -26,87 +27,58 @@ const HomepageAbout = () => {
       </div>
       <div className="container">
         <Swiper {...params} getSwiper={setSwiper}>
-          <div className="heading__container">
-            <div className="homepage__about-left">
-              <h2 className="heading__title mt-5 mb-5 tw-text-blue-900">
-                &lsaquo;ABOUT_US/&rsaquo;
-              </h2>
-              <p className="lead tw-text-blue-900">
-                We are an international network of developers unifying together
+          <span>
+            <AboutSectionItem
+              headingTitle="ABOUT_US"
+              description="We are an international network of developers unifying together
                 to build socially impactful projects &amp; spread STEM education
-                to marginalized communities.
-              </p>
-            </div>
-            <div className="homepage__about-right">
-              <img src="/assets/images/globe.svg" className="heading__img" alt="" />
-              <button name="play" onClick={() => setOpen(true)} />
-            </div>
-          </div>
-          <div className="heading__container">
-            <div className="homepage__about-left">
-              <h2 className="heading__title mt-5 mb-5 tw-text-blue-900">
-                &lsaquo;OUR_PEOPLE/&rsaquo;
-              </h2>
-              <p className="lead tw-text-blue-900">
-                We are a place for minorities to come meet friends, gain
+                to marginalized communities."
+              imgSrcList={['/assets/images/globe.svg']}
+              onPlayBtnClick={() => setOpen(true)}
+            />
+          </span>
+          <span>
+            <AboutSectionItem
+              headingTitle="OUR_PEOPLE/"
+              description="We are a place for minorities to come meet friends, gain
                 guidance, and learn relevant IT skills to become a world leader
-                in tech.
-              </p>
-            </div>
-            <div className="homepage__about-right resize">
-              <img src="/assets/images/about2.2.svg" className="heading__img" alt="" />
-              <img src="/assets/images/about2.1.svg" className="heading__img" alt="" />
-              <button name="play" onClick={() => setOpen(true)} />
-            </div>
-          </div>
-          <div className="heading__container">
-            <div className="homepage__about-left">
-              <h2 className="heading__title mt-5 mb-5 tw-text-blue-900">
-                &lsaquo;OUR_VISION/&rsaquo;
-              </h2>
-              <p className="lead tw-text-blue-900">
-                Complete turnkey decentralized vertical integration for scaling
+                in tech."
+              imgBoxCustomClassNames="resize"
+              imgSrcList={['/assets/images/about2.2.svg', '/assets/images/about2.1.svg']}
+              onPlayBtnClick={() => setOpen(true)}
+            />
+          </span>
+          <span>
+            <AboutSectionItem
+              headingTitle="OUR_VISION/"
+              description="Complete turnkey decentralized vertical integration for scaling
                 minority startups from ideation to large scale businesses using
-                decentralized protocols.
-              </p>
-            </div>
-            <div className="homepage__about-right">
-              <img src="/assets/images/about4.svg" className="heading__img" alt="" />
-              <button name="play" onClick={() => setOpen(true)} />
-            </div>
-          </div>
-          <div className="heading__container">
-            <div className="homepage__about-left">
-              <h2 className="heading__title mt-5 mb-5 tw-text-blue-900">
-                &lsaquo;FOR_COMPANIES/&rsaquo;
-              </h2>
-              <p className="lead tw-text-blue-900">
-                We are a one-stop-shop for companies to recruit, train, and
+                decentralized protocols."
+              imgSrcList={['/assets/images/about4.svg']}
+              onPlayBtnClick={() => setOpen(true)}
+            />
+          </span>
+          <span>
+            <AboutSectionItem
+              headingTitle="FOR_COMPANIES/"
+              description="We are a one-stop-shop for companies to recruit, train, and
                 onboard diverse talent that cares about advancing the community
-                through technology
-              </p>
-            </div>
-            <div className="homepage__about-right">
-              <img src="/assets/images/about5.svg" className="heading__img" alt="" />
-              <button name="play" onClick={() => setOpen(true)} />
-            </div>
-          </div>
-          <div className="heading__container">
-            <div className="homepage__about-left">
-              <h2 className="heading__title mt-5 mb-5 tw-text-blue-900">
-                &lsaquo;OUR_APP/&rsaquo;
-              </h2>
-              <p className="lead tw-text-blue-900">
-                We are a lifelong community run platform that gives minorities
+                through technology"
+              imgSrcList={['/assets/images/about5.svg']}
+              onPlayBtnClick={() => setOpen(true)}
+            />
+          </span>
+          <span>
+            <AboutSectionItem
+              headingTitle="OUR_APP/"
+              description="We are a lifelong community run platform that gives minorities
                 access to opportunities via ideas, resources, mentorship,
-                education, and access to capital
-              </p>
-            </div>
-            <div className="homepage__about-right resize">
-              <img src="/assets/images/about3.svg" className="heading__img" alt="" />
-              <button name="play" onClick={() => setOpen(true)} />
-            </div>
-          </div>
+                education, and access to capital"
+              imgBoxCustomClassNames="resize"
+              imgSrcList={['/assets/images/about3.svg']}
+              onPlayBtnClick={() => setOpen(true)}
+            />
+          </span>
         </Swiper>
         <ModalVideo
           channel="youtube"
