+++ components/homepage/HomepageCore.js
@@ -1,128 +1,7 @@
 import React, { useState } from 'react';
 import Swiper from 'react-id-swiper';
 import 'swiper/css/swiper.css';
-
-const CORE_PRINCIPLES = [
-  {
-    id: '1',
-    image: '/assets/images/core1.png',
-    icon: '/assets/images/core1.svg',
-    color: 'pink',
-    title: 'Diversity in stem',
-    description:
-      'Our priority for diversity and collaboration, has urged us to create an environment that fosters inclusivity and diversity of both cultures and perspective',
-    url: '#',
-  },
-  {
-    id: '2',
-    image: '/assets/images/core2.png',
-    icon: '/assets/images/core2.svg',
-    color: 'blue',
-    title: 'Software engineer',
-    description: 'We focus on shipping out industry grade software products',
-    url: '#',
-  },
-  {
-    id: '3',
-    image: '/assets/images/core3.png',
-    icon: '/assets/images/core3.svg',
-    color: 'yellow',
-    title: 'Entrepreneuship',
-    description:
-      'We ignite entrepreneurship within minority communities as an avenue for innovation and upward mobility',
-    url: '#',
-  },
-  {
-    id: '4',
-    image: '/assets/images/core4.png',
-    icon: '/assets/images/core4.svg',
-    color: 'pink',
-    title: 'Interdisciplinary learning',
-    description:
-      ' We motivate ourselves to learn by making connections between ideas and concepts across different professional and academic disciplines.',
-    url: '#',
-  },
-  {
-    id: '5',
-    image: '/assets/images/core5.png',
-    icon: '/assets/images/core5.svg',
-    color: 'blue',
-    title: 'Innovations',
-    description:
-      'We nurture new ideas and encourage our employees to think out of the box. There is no one solution to all problems',
-    url: '#',
-  },
-  {
-    id: '6',
-    image: '/assets/images/core6.png',
-    icon: '/assets/images/core6.svg',
-    color: 'yellow',
-    title: 'Project-based learning',
-    description:
-      'We believe in project based learning as the most effective means of delivering professional team driven experiences to students around the world',
-    url: '#',
-  },
-  {
-    id: '7',
-    image: '/assets/images/core7.png',
-    icon: '/assets/images/core7.svg',
-    color: 'pink',
-    title: 'Professional development',
-    description:
-      'We provide resources for continual learning and skills enhancement to adapt to the changes in progressive technological advancements.',
-    url: '#',
-  },
-  {
-    id: '8',
-    image: '/assets/images/core8.png',
-    icon: '/assets/images/core8.svg',
-    color: 'blue',
-    title: 'activating passions',
-    description:
-      'Passion activates purpose. We strongly believe in creating a community that fuels your passion and ideas.',
-    url: '#',
-  },
-  {
-    id: '9',
-    image: '/assets/images/core9.png',
-    icon: '/assets/images/core9.svg',
-    color: 'yellow',
-    title: 'Sense of community',
-    description:
-      'We strive to build a supportive community for developers to engage, share knowledge, motivate and grow into valuable assets.',
-    url: '#',
-  },
-  {
-    id: '10',
-    image: '/assets/images/core10.png',
-    icon: '/assets/images/core10.svg',
-    color: 'pink',
-    title: 'Training/onboarding',
-    description:
-      'Our training and onboarding has been meticulously designed to focus on new employees social and performance',
-    url: '#',
-  },
-  {
-    id: '11',
-    image: '/assets/images/core11.png',
-    icon: '/assets/images/core11.svg',
-    color: 'blue',
-    title: 'Humanitarian focused',
-    description:
-      'We build products that at its core help people improve humanity',
-    url: '#',
-  },
-  {
-    id: '12',
-    image: '/assets/images/core11.png',
-    icon: '/assets/images/core11.svg',
-    color: 'yellow',
-    title: 'Mentorship',
-    description:
-      ' Our mentorship program provides learning opportunities, expands your professional network and to gain new perspectives to build a successful career.',
-    url: '#',
-  },
-];
+import CORE_PRINCIPLES from './corePrinciplesData';
 
 const HomepageCore = () => {
   const [popupInfo, setPopupInfo] = useState({
