+++ components/homepage/HomepageTestimonials.js
@@ -1,112 +1,9 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import Swiper from 'react-id-swiper';
 import 'swiper/css/swiper.css';
+import { allies, sponsors, TestimonialsCollection } from './testimonialSectionData';
 
 const HomepageTestimonials = () => {
-  const TestimonialsCollection = [
-    {
-      id: 1,
-      name: 'Bryanna Turman',
-      position: 'Social Media Manager',
-      quote:
-        'MPA hits different. Its a whole gang of coders trying to put on for the world.',
-      imgSrc: 'https://minorityprogrammers.com/assets/images/Bryanna.svg',
-    },
-    {
-      id: 2,
-      name: 'Seemaab Mujtaba',
-      position: 'Software Engineer Intern',
-      quote:
-        'MPA is the most fun work experience I had and I always had a place to express my ideas!',
-      imgSrc: 'https://minorityprogrammers.com/assets/images/Seemaab.svg',
-    },
-  ];
-
-  const sponsors = [
-    {
-      sponsorLink: 'https://blockchainedu.org/',
-      imgSrc: 'assets/images/cgi.png',
-    },
-    {
-      sponsorLink: 'https://www.accenture.com/us-en',
-      imgSrc: 'assets/images/Accenture-Logo.png',
-    },
-    {
-      sponsorLink: 'https://www.captechconsulting.com/',
-      imgSrc: 'assets/images/caotech.png',
-    },
-    {
-      sponsorLink: 'https://enzyme.finance/',
-      imgSrc: 'assets/images/enzyme-logo.png',
-    },
-    {
-      sponsorLink: 'https://www.stellar.org/',
-      imgSrc: 'assets/images/stellar-logo.png',
-    },
-    {
-      sponsorLink: 'https://www.plutopepe.com/',
-      imgSrc: 'assets/images/pluto-logo.png',
-    },
-    {
-      sponsorLink: 'https://ripplex.io/',
-      imgSrc: 'assets/images/sponsorship/sponsor9.png',
-    },
-    {
-      sponsorLink: 'https://idle.finance/',
-      imgSrc: 'assets/images/sponsorship/sponsor7.png',
-    },
-    {
-      sponsorLink: 'https://tokamak.network/',
-      imgSrc: 'assets/images/sponsorship/sponsor8.png',
-    },
-  ];
-
-  const allies = [
-    {
-      allyLink: 'https://mason360.gmu.edu/bam/',
-      imgSrc: 'assets/images/brand-1-1.png',
-    },
-    {
-      allyLink: 'https://thelastmile.org/',
-      imgSrc: 'assets/images/lastmile.png',
-    },
-    { allyLink: 'https://aisnet.org/', imgSrc: 'assets/images/ais.png' },
-    {
-      allyLink: 'https://www.blockchainedu.org/',
-      imgSrc: 'assets/images/ben.png',
-    },
-    { allyLink: 'https://devpost.com/', imgSrc: 'assets/images/devpost.png' },
-    {
-      allyLink: 'https://www.blockchainacceleration.org/',
-      imgSrc: 'assets/images/baf-logo.png',
-    },
-    { allyLink: 'https://www.jmu.edu/', imgSrc: 'assets/images/jmucs.png' },
-    {
-      allyLink: 'https://www.echelondigital.com/',
-      imgSrc: 'assets/images/echelon-logo.png',
-    },
-    {
-      allyLink: 'https://montech.io/',
-      imgSrc: 'assets/images/sponsorship/sponsor10.png',
-    },
-    {
-      allyLink: 'https://jmuxlabs.org/',
-      imgSrc: 'assets/images/sponsorship/sponsor11.png',
-    },
-    {
-      allyLink: 'https://www.learnacademy.org/',
-      imgSrc: 'assets/images/sponsorship/sponsorlearn.svg',
-    },
-    {
-      allyLink: 'https://algorand.foundation/',
-      imgSrc: 'assets/images/sponsorship/sponsor13.png',
-    },
-    {
-      allyLink: 'https://entethalliance.org/',
-      imgSrc: 'assets/images/sponsorship/sponsor14.png',
-    },
-  ];
-
   const [swiper, setSwiper] = useState(null);
   const [currentIndex, updateCurrentIndex] = useState(0);
